+++ Assembler/my_assembler.h
@@ -3,7 +3,7 @@
 #include <stdio.h>
 #include "text.h"
 
-#define ERROR 0xDED00DED
+#define ERROR (int)0xDED00DED
 #define NUM_OF_COMMANDS 25
 #define NUM_OF_LINKS 50
 #define REGISTER_COUNT 6
